refactor(server): isolate chat read model
