+++ packages/server/src/services/session.ts
@@ -1,4 +1,4 @@
-import { MENTION_REGEX, type SessionState, stripCode } from "@first-tree/shared";
+import type { SessionState } from "@first-tree/shared";
 import { and, desc, eq, inArray, lt, ne, sql } from "drizzle-orm";
 import type { Database } from "../db/connection.js";
 import { agentChatSessions } from "../db/schema/agent-chat-sessions.js";
@@ -10,45 +10,10 @@ import { inboxEntries } from "../db/schema/inbox-entries.js";
 import { messages } from "../db/schema/messages.js";
 import { sessionEvents } from "../db/schema/session-events.js";
 import { BadRequestError, ConflictError, NotFoundError } from "../errors.js";
+import { extractChatSummary } from "./chat-read-model.js";
 import type { Notifier } from "./notifier.js";
 import { agentRouteGuardSql } from "./session-command-rpc.js";
 import * as sessionEventService from "./session-event.js";
-
-export const SUMMARY_MAX_LENGTH = 50;
-
-/** Extract a plain-text summary from a message's JSONB content field.
- *  Used as the auto-title fallback in chat list rendering — see
- *  `me-chat.ts:resolveChatTitle` and `admin/chats.ts:getChat`.
- *
- *  - `@<name>` mention tokens are stripped before truncation: in the
- *    chat-first model they're routing/audience metadata, not part of
- *    the user's intent. Leaving them in produces noisy titles like
- *    "@agent-01 帮我重构这个文件" or "你好 @agent-02 看看".
- *  - Whitespace runs (including those left behind by mention removal)
- *    collapse to single spaces.
- *  - If the cleaned text is empty (e.g., a message that's only
- *    `@agent-01`), returns null so the caller falls through to
- *    the participant-join fallback.
- *  - Slicing is code-point-aware (`Array.from + join`) so emoji /
- *    surrogate pairs aren't split into garbled half-characters. */
-export function extractSummary(content: unknown, maxLen = SUMMARY_MAX_LENGTH): string | null {
-  let text = "";
-  if (typeof content === "object" && content !== null && "text" in content) {
-    text = String((content as { text: unknown }).text ?? "");
-  } else if (typeof content === "string") {
-    text = content;
-  }
-  if (!text) return null;
-  // `stripCode` first so identifier-shaped tokens inside Markdown
-  // code regions (`` `@param` ``, fenced blocks) aren't misclassified
-  // as mentions and stripped — that would produce titles like
-  // `"Use  decorator"` from `"Use \`@param\` decorator"`. Mirrors
-  // `extractMentions`'s pipeline so routing and titling agree on what
-  // counts as a real mention vs a code reference.
-  const cleaned = stripCode(text).replace(MENTION_REGEX, "").replace(/\s+/g, " ").trim();
-  if (!cleaned) return null;
-  return Array.from(cleaned).slice(0, maxLen).join("");
-}
 
 export type SessionListItem = {
   agentId: string;
@@ -200,7 +165,7 @@ export async function listAgentSessions(
 
   const summaryMap = new Map<string, string>();
   for (const row of firstMessages) {
-    const summary = extractSummary(row.content);
+    const summary = extractChatSummary(row.content);
     if (summary) {
       summaryMap.set(row.chatId, summary);
     }
@@ -259,7 +224,7 @@ export async function getSession(db: Database, agentId: string, chatId: string):
     sql`SELECT content FROM messages WHERE chat_id = ${chatId} ORDER BY created_at ASC LIMIT 1`,
   );
   const firstMsg = firstMsgRows[0];
-  const summary = firstMsg ? extractSummary(firstMsg.content) : null;
+  const summary = firstMsg ? extractChatSummary(firstMsg.content) : null;
 
   return {
     agentId: row.agentId,
@@ -363,7 +328,7 @@ export async function listAllSessions(
       : [];
   const summaryMap = new Map<string, string>();
   for (const row of firstMessages) {
-    const summary = extractSummary(row.content);
+    const summary = extractChatSummary(row.content);
     if (summary) summaryMap.set(row.chatId, summary);
   }
